🔨 Tidy `watch.sh`
I haven't had cargo bench run in this in a while. It takes too long.

A nice TODO would be to write a script that runs the benchmarks, and
then packages the output data and graphs somewhere safe. I run
`cargo clean` *way* too much to leave valuable data in `target/`.